fix(ui): correct color integer conversion in HexCanvas
Change the color rendering logic to use toInt() instead of toULong() to ensure proper color value interpretation for the hex tiles.

+++ app/src/main/java/com/example/ui/HexCanvasScreen.kt
@@ -178,7 +178,7 @@ fun HexCanvasScreen(
                                 
                                 translate(center.x, center.y) {
                                     if (userHex != null) {
-                                        drawPath(basePath, Color(userHex.color.toULong()))
+                                        drawPath(basePath, Color(userHex.color.toInt()))
                                     } else {
                                         drawPath(basePath, borderColor, style = Stroke(width = 1f))
                                     }
